+++ pkg/azurecluster/mutate_create.go
@@ -124,14 +124,6 @@ func (m *CreateMutator) Mutate(ctx context.Context, request *v1beta1.AdmissionRe
 		result = append(result, *patch)
 	}
 
-	patch, err = m.ensureControlPlaneSubnet(ctx, azureClusterCR)
-	if err != nil {
-		return []mutator.PatchOperation{}, microerror.Mask(err)
-	}
-	if patch != nil {
-		result = append(result, *patch)
-	}
-
 	patch, err = generic.EnsureComponentVersionLabelFromRelease(ctx, m.ctrlClient, azureClusterCR.GetObjectMeta(), "azure-operator", label.AzureOperatorVersion)
 	if err != nil {
 		return []mutator.PatchOperation{}, microerror.Mask(err)
@@ -220,28 +212,6 @@ func (m *CreateMutator) ensureAPIServerLBFrontendIPs(ctx context.Context, azureC
 		}
 
 		return mutator.PatchAdd("/spec/networkSpec/apiServerLB/frontendIPs", frontendIPs), nil
-	}
-
-	return nil, nil
-}
-
-func (m *CreateMutator) ensureControlPlaneSubnet(ctx context.Context, azureCluster *capzv1alpha3.AzureCluster) (*mutator.PatchOperation, error) {
-	hasControlPlaneSubnet := false
-	for _, subnet := range azureCluster.Spec.NetworkSpec.Subnets {
-		if subnet.Role == capzv1alpha3.SubnetControlPlane {
-			hasControlPlaneSubnet = true
-			break
-		}
-	}
-
-	if !hasControlPlaneSubnet {
-		subnets := azureCluster.Spec.NetworkSpec.Subnets[:]
-		subnets = append(subnets, &capzv1alpha3.SubnetSpec{
-			Role: capzv1alpha3.SubnetControlPlane,
-			Name: key.MasterSubnetName(azureCluster.Name),
-		})
-
-		return mutator.PatchAdd("/spec/networkSpec/subnets", subnets), nil
 	}
 
 	return nil, nil